Created Queries and Test traits/classes
Going to use Queries to store the expected results (maybe I should move
it into the test folder) and created a trait for Booting and closing the
ES node.

Need to maybe write a trait for loading some data or something if I want
to write tests that do things with actual data.